Upload
The actual code works properly, all the json representations of the
objects will be saved in the file '/models/storage.json', the items in the
file can be updated, there is no way to modify the actual path without
touch the code. the Default values of ´´name´´ and ´´my_number´´ are an
empty string and 0 repectively.

The code has documentation and no trouble with the pep8 style